Add handleSkogbrukWMSFeatures function to CustomMapEvents.js and remove duplicate function from Map.js
